usmle_error_note/requirements.txt
crewai
crewai-tools
fastapi
uvicorn[standard]
pyyaml
python-dotenv
